+++ src/lib/net/iTin.Hardware/iTin.Hardware.Specification/iTin.Hardware.Specification.Eedid/EEDID/Blocks/DisplayID/Sections/DataBlocks/ComponentModel/DataBlockFactory.cs
@@ -1,0 +1,21 @@
+﻿
+namespace iTin.Hardware.Specification.Eedid.Blocks.DisplayId.Sections.DataBlocks.ComponentModel
+{
+    using System;
+
+    internal static class DataBlockFactory
+    {
+        public static BaseDataSection GetDataBlockData(DataBlockData data)
+        {
+            var type = data.BlockTag;
+            switch (type)
+            {
+                case DataBlockTag.VendorSpecific:
+                    return new VendorSpecificDataBlock(data.RawData, data.StructureVersion);
+
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+        }
+    }
+}